♻️ refactor:  全体的に構成を見直し
#65
